+++ contracts/interfaces/ICoverPoolFactory.sol
@@ -24,16 +24,16 @@ abstract contract ICoverPoolFactory {
     function createCoverPool(
         address fromToken,
         address destToken,
-        uint256 fee,
-        uint24  tickSpread,
+        uint16  fee,
+        int16   tickSpread,
         uint16  twapLength
     ) external virtual returns (address book);
 
     function getCoverPool(
         address fromToken,
         address destToken,
-        uint256 fee,
-        uint24  tickSpread,
+        uint16 fee,
+        int16  tickSpread,
         uint16  twapLength
     ) external virtual view returns (address);
 }
