+++ typechain/factories/CoverPoolStorage__factory.ts
@@ -108,14 +108,14 @@ const _abi = [
         type: "uint8",
       },
       {
-        internalType: "uint24",
+        internalType: "uint16",
         name: "swapFee",
-        type: "uint24",
+        type: "uint16",
       },
       {
-        internalType: "int24",
+        internalType: "int16",
         name: "tickSpread",
-        type: "int24",
+        type: "int16",
       },
       {
         internalType: "uint16",
@@ -330,14 +330,14 @@ const _abi = [
     name: "ticks0",
     outputs: [
       {
-        internalType: "int104",
+        internalType: "int128",
         name: "liquidityDelta",
-        type: "int104",
+        type: "int128",
       },
       {
-        internalType: "uint104",
+        internalType: "uint128",
         name: "liquidityDeltaMinus",
-        type: "uint104",
+        type: "uint128",
       },
       {
         internalType: "uint128",
@@ -345,14 +345,14 @@ const _abi = [
         type: "uint128",
       },
       {
-        internalType: "int88",
+        internalType: "int128",
         name: "amountInDelta",
-        type: "int88",
+        type: "int128",
       },
       {
-        internalType: "int88",
+        internalType: "int128",
         name: "amountOutDelta",
-        type: "int88",
+        type: "int128",
       },
       {
         internalType: "uint64",
@@ -379,14 +379,14 @@ const _abi = [
     name: "ticks1",
     outputs: [
       {
-        internalType: "int104",
+        internalType: "int128",
         name: "liquidityDelta",
-        type: "int104",
+        type: "int128",
       },
       {
-        internalType: "uint104",
+        internalType: "uint128",
         name: "liquidityDeltaMinus",
-        type: "uint104",
+        type: "uint128",
       },
       {
         internalType: "uint128",
@@ -394,14 +394,14 @@ const _abi = [
         type: "uint128",
       },
       {
-        internalType: "int88",
+        internalType: "int128",
         name: "amountInDelta",
-        type: "int88",
+        type: "int128",
       },
       {
-        internalType: "int88",
+        internalType: "int128",
         name: "amountOutDelta",
-        type: "int88",
+        type: "int128",
       },
       {
         internalType: "uint64",
